[chore][pkg/stanza] Cleanup keyvalue parser operator files
