+++ pkg/stanza/operator/parser/keyvalue/parser_test.go
